RHIInit移动到Render前存档

+++ AntiTitanEngine/AntiTitanEngine/ActorsInfo.cpp
@@ -29,7 +29,7 @@ void ActorsInfo::SetSceneActorsInfoFromBat(const std::string& StaticMeshPath) {
 		ActorNameArray[i].resize(DataStringLength);
 		BatFile.read((char*)ActorNameArray[i].data(), sizeof(char) * DataStringLength);
 	}
-	pos.push_back(BatFile.tellg());
+	pos.push_back(static_cast<unsigned int>(BatFile.tellg()));
 
 	//MeshNameArray
 	BatFile.read((char*)&DataLength, sizeof(int32_t));
@@ -40,21 +40,19 @@ void ActorsInfo::SetSceneActorsInfoFromBat(const std::string& StaticMeshPath) {
 		MeshNameArray[i].resize(DataStringLength);
 		BatFile.read((char*)MeshNameArray[i].data(), sizeof(char) * DataStringLength);
 	}
-	pos.push_back(BatFile.tellg());
-
+	pos.push_back(static_cast<unsigned int>(BatFile.tellg()));
 
 	//ActorsTransformArray
 	BatFile.read((char*)&DataLength, sizeof(int32_t));
 	ActorsTransformArray.resize(DataLength);
 	BatFile.read((char*)ActorsTransformArray.data(), sizeof(Transform) * DataLength);
-	pos.push_back(BatFile.tellg());
-
+	pos.push_back(static_cast<unsigned int>(BatFile.tellg()));
 
 	//ActorsQuatArray
 	BatFile.read((char*)&DataLength, sizeof(int32_t));
 	ActorsQuatArray.resize(DataLength);
 	BatFile.read((char*)ActorsQuatArray.data(), sizeof(Quat) * DataLength);
-	pos.push_back(BatFile.tellg());
+	pos.push_back(static_cast<unsigned int>(BatFile.tellg()));
 
 	BatFile.close();
 }
